% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_activity_single.R
\name{LDA_activity_single}
\alias{LDA_activity_single}
\title{LDA_activity_single}
\usage{
LDA_activity_single(x, name = "cell line a", treat = "no")
}
\arguments{
\item{x}{numeric data.frame or matrix with three columns (dose, number of
wells, number of positive wells)}

\item{name}{optional: experiment name (e.g. name of cell line)}

\item{xtreat}{optional: treatment}
}
\value{
list object with estimated activity, 95%-confidence interval,
  84%-confidence interval, estimated parameters, corresponding covariance
  matrix, fit-object and p-value for non-cooperativity-test
}
\description{
calculation of activity in a block of LDA data
  (i.e. dose, number of wells, number of positive wells).
}
\examples{
x <- data.frame("dose" = c(10,50,100,250),
                "wells" = rep(25,4),
                "positive" = c(2,5,10,20))
act <- LDA_activity_single(x)
}
