% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plot.R
\name{LDA_plot}
\alias{LDA_plot}
\title{LDA_plot}
\usage{
LDA_plot(LDA_tab, uncertainty = "act", xlim = NULL, uncertainty.band = FALSE)
}
\arguments{
\item{LDA_tab}{LDA data.frame
("cells", "wells", "positive", "group", "replicate")}
}
\value{
none
}
\description{
plot activity from LDA data
}
\examples{
data(LDAdata)
Z1 <- subset.data.frame(LDAdata,subset = name == unique(LDAdata$name)[1])
LDA_plot(Z1[,c("S-value","# Tested","# Clonal growth","Group","replicate")])
}
