% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plot_SF.R
\name{LDA_plot_SF}
\alias{LDA_plot_SF}
\title{LDA_plot_SF}
\usage{
LDA_plot_SF(LDA_obj)
}
\arguments{
\item{LDA_obj}{list returned from LDA_prepare_plot}
}
\value{
none
}
\description{
generate clonogenic survival plot (estimated clonogenic
  survival over treatment) for data from limiting dilution assay (LDA).
  Input is an data object as returned by the preprocessing function
  LDA_prepare_plot().
}
\examples{
x <- data.frame("cells" = rep(c(10,50,100,250),times = 4),
                "wells" = rep(25,16),
                "positive" = c(2,5,10,20,1,2,6,11,3,4,8,22,1,1,7,12),
                "group" = rep(c(rep(0,4),rep(6,4)),times = 2),
                "replicate" = c(rep(1,8),rep(2,8)))
out <- LDA_prepare_plot(x)
LDA_plot_SF(out[[2]])
data(LDAdata)
Z1 <- subset.data.frame(LDAdata,subset = name == unique(LDAdata$name)[1])
out <- LDA_prepare_plot(Z1[,c("S-value","# Tested","# Clonal growth",
                                "Group","replicate")])
LDA_plot_SF(out[[2]])
}
