% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_survival_single.R
\name{LDA_survival_single}
\alias{LDA_survival_single}
\title{LDA_suvival_single}
\usage{
LDA_survival_single(act.0, act.x)
}
\arguments{
\item{act.0}{reference activity}

\item{act.x}{activity after treatment}
}
\value{
list object with survival fraction, estimated confidence intervals
  (by error propagation through first order Taylor series approximation and
  by combination of 84%-uncertainty-intervals of activity estimates)
}
\description{
calculate clonogenic survival fraction from LDA_activity objects
}
\examples{
x <- data.frame("dose" = c(10,50,100,250,10,50,100,250),
                "wells" = rep(25,8),
                "positive" = c(2,5,10,20,1,2,6,11),
                "group" = c(rep("A",4),rep("B",4)))
act <- LDA_activity(x)
sf <- LDA_survival(act)
}
