% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_table.R
\name{LDA_table}
\alias{LDA_table}
\title{LDA_table}
\usage{
LDA_table(x, ref_class = "unknown")
}
\arguments{
\item{x}{numeric data.frame or matrix with at least three columns (dose,
number of wells, number of positive wells, group (optional))}
}
\value{
table
}
\description{
show table with activities and clonogenic survival from LDA data
  object
}
\examples{
x <- data.frame("dose" = c(10,50,100,250,10,50,100,250),
                "wells" = rep(25,8),
                "positive" = c(2,5,10,20,1,2,6,11),
                "group" = c(rep("A",4),rep("B",4)))
LDA_table(x)
}
