% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_plot.R
\name{LDA_plot}
\alias{LDA_plot}
\title{LDA_plot}
\usage{
LDA_plot(LDA_act)
}
\arguments{
\item{LDA_act}{output from LDA_activity()}
}
\value{
none
}
\description{
plot activity from LDA data object
}
\examples{
x <- data.frame("dose" = c(10,50,100,250,10,50,100,250),
                "wells" = rep(25,8),
                "positive" = c(2,5,10,20,1,2,6,11),
                "group" = c(rep("A",4),rep("B",4)))
LDA_plot(LDA_activity(x))

}
