% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAdata.R
\docType{data}
\name{LDAdata}
\alias{LDAdata}
\title{LDA (limiting dilution assay) data from a set of cell lines}
\format{
data.frame with columns: "name", "replicate", "Group", "S-value",
"# Tested", "# Clonal growth"
}
\usage{
data(LDAdata)
}
\description{
LDA data from 11 cell lines, up to 4 biological replicates and up to 6
treatments.
}
\examples{
data(LDAdata)
}
\keyword{datasets}
